+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Pass full BranchUpdate result on successful merge
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/406
 - Deprecate implementation of RepositoryService.Exists
   https://gitlab.com/gitlab-org/gitaly/merge_requests/408
 - Use gitaly-proto 0.42.0